Fixing button issues
at the create of this branch, the list of file paths the program referenced when it was checking what image needed to be copied to the clipboard when you clicked a button would only be the list of images for the final section created when the window was populated. This resulted in a crash if a button in a section was clicked that was in a greater position within that section that the total count of buttons in the final section created. This also resulted in the nth button in each section copying the same image, which again would be the same image as the nth button in the final section created. Basically, each section because equivalent to the final section in terms of working number of button and the images it copied to your clipboard.